Use function prototypes & move config.h inclusion
Instead of just sticking '#include "config.h"' after all the functions
it needs to use, clarify what is going on with function prototypes and
keep all #includes in the top 24 lines.

Reminder: config.h is not a proper header file, and should include
functions and data, similar in spirit to dwm's config.h.